docs(schema): Add edition_title + more validation
Add the new field `edition_title` inline with the change to the Tier 0
metadata model.

Improve the validation indicators by including minLength for fields that
can not be empty, specifying that at least one distribution file is
needed and indicating that only @ons.gov.uk and @ext.ons.gov.uk emails are
allowed for the pipeline notifications.